Update Engine mode to be int
The C standard says that an enum is not defined to be an unsigned or signed
value. While running the test suite under Go 1.11.4 there were errors about
CS_MODE_BIG_ENDIAN overflowing the uint value. It seems similar to this:

https://github.com/golang/go/issues/24523

Since the mapping of Capstone constants to Go constants is untyped it means
that the values are of type int. To solve this I changed the engine mode to be of
type int.

I think another potential way to solve this is to explicitly specify a type for
the engine mode constants.